# More import time improvements.