--- a/tests/compiled/test_import_compile.py
+++ b/tests/compiled/test_import_compile.py
@@ -1,6 +1,6 @@
 import pytest
 
-from prefab_classes import prefab_compiler
+from prefab_classes.compiled import prefab_compiler
 from prefab_classes.exceptions import CompiledPrefabError