--- a/tests/conftest.py
+++ b/tests/conftest.py
@@ -1,6 +1,6 @@
 import pytest
 
-from prefab_classes import prefab_compiler
+from prefab_classes.compiled import prefab_compiler
 
 
 @pytest.fixture(params=[True, False], ids=["Compiled", "Live"])