--- a/tests/shared/test_creation.py
+++ b/tests/shared/test_creation.py
@@ -1,5 +1,5 @@
 """Tests for errors raised on class creation"""
-from prefab_classes import PrefabError
+from prefab_classes.exceptions import PrefabError
 from prefab_classes.constants import FIELDS_ATTRIBUTE
 
 import pytest