--- a/tests/shared/test_serialization.py
+++ b/tests/shared/test_serialization.py
@@ -1,7 +1,7 @@
 """Tests related to serialization to JSON or Pickle"""
 from pathlib import PurePosixPath, PurePath
 
-from prefab_classes import as_dict, to_json
+from prefab_classes.funcs import as_dict, to_json
 from pytest import raises